Fix public games on mobile
