+++ src/app/_components/HomePage/_subcomponents/PublicMatch/PublicMatch.tsx
@@ -1,52 +1,38 @@
 import React from 'react';
-import { Box, Button, Typography, Popover, PopoverOrigin } from '@mui/material';
-import GameInProgressPlayer from '../GameInProgressPlayer/GameInProgressPlayer';
+import { Box, Button, Typography } from '@mui/material';
 import { IPublicGameInProgressProps } from '../../HomePageTypes';
-import { s3CardImageURL } from '@/app/_utils/s3Utils';
-import { useCardImageLocale } from '@/app/_contexts/CardImageLocale.context';
 import { useRouter } from 'next/navigation';
 import { useUser } from '@/app/_contexts/User.context';
-import { LeaderBaseCardStyle } from '@/app/_components/_sharedcomponents/Cards/CardTypes';
+import OverlappingCards from '../OverlappingCards/OverlappingCards';
+
+const styles = {
+    box: {
+        width: '100%',
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        mb: '2.5rem',
+        pl: '10px',
+        gap: '1rem',
+    },
+    matchItems: {
+        display: 'flex',
+        alignItems: 'center',
+        gap: '1rem',
+        flex: 1,
+    },
+    matchType: {
+        margin: 0,
+        mr: '1rem',
+    },
+    overlappingCard: {
+        width: 'clamp(5rem, 100%, 14rem)', // Min 5rem, max 10rem, scales with viewport
+    }
+};
 
 const PublicMatch: React.FC<IPublicGameInProgressProps> = ({ match }) => {
     const router = useRouter();
     const { user } = useUser();
-    const locale = useCardImageLocale();
-    const [anchorElement, setAnchorElement] = React.useState<HTMLElement | null>(null);
-    const [previewImage, setPreviewImage] = React.useState<string | null>(null);
-    const hoverTimeout = React.useRef<number | undefined>(undefined);
-    const open = Boolean(anchorElement);
-
-    const handlePreviewOpen = (event: React.MouseEvent<HTMLElement>) => {
-        const target = event.currentTarget;
-        const imageUrl = target.getAttribute('data-card-url');
-
-        if (!imageUrl) return;
-
-        hoverTimeout.current = window.setTimeout(() => {
-            setAnchorElement(target);
-            setPreviewImage(`url(${imageUrl})`);
-        }, 300);
-    };
-
-    const handlePreviewClose = () => {
-        clearTimeout(hoverTimeout.current);
-        setAnchorElement(null);
-        setPreviewImage(null);
-    };
-
-    const popoverConfig = (): { anchorOrigin: PopoverOrigin, transformOrigin: PopoverOrigin } => {
-        return {
-            anchorOrigin: {
-                vertical: 'center',
-                horizontal: 'right',
-            },
-            transformOrigin: {
-                vertical: 'center',
-                horizontal: 'left',
-            }
-        };
-    };
 
     const handleSpectate = () => {
         router.push(`/spectate?lobbyId=${match.id}`);
@@ -54,115 +40,28 @@ const PublicMatch: React.FC<IPublicGameInProgressProps> = ({ match }) => {
 
     const spectateDisabled = !user && process.env.NODE_ENV !== 'development';
 
-    const styles = {
-        cardPopover: {
-            borderRadius: '.38em',
-            backgroundSize: 'cover',
-            backgroundRepeat: 'no-repeat',
-            aspectRatio: '1.4 / 1',
-            width: '24rem',
-        },
-        box: {
-            width: '100%',
-            display: 'flex',
-            justifyContent: 'space-between',
-            alignItems: 'center',
-            mb: '2.5rem',
-            pl: '10px',
-        },
-        matchItems: {
-            display: 'flex',
-            alignItems: 'center',
-            gap: '1rem',
-        },
-        matchType: {
-            margin: 0,
-            mr: '1rem',
-        },
-        leaderStyleCard:{
-            borderRadius: '0.5rem',
-            backgroundSize: 'cover',
-            width: 'clamp(3rem, 7vw, 10rem)', // Min 5rem, max 10rem, scales with viewport
-            aspectRatio: '1.39',
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            position: 'relative',
-            border: '2px solid transparent',
-            boxSizing: 'border-box',
-            cursor: 'pointer'
-        },
-        parentBoxStyling: {
-            position:'absolute',
-        },
-    };
-
     return (
-        <>
-            <Box sx={styles.box}>
-                <Box sx={styles.matchItems}>
-                    <Box sx={{ position:'relative' }}>
-                        <Box>
-                            <Box 
-                                sx={{ ...styles.leaderStyleCard,backgroundImage:`url(${s3CardImageURL(match.player1Base, locale)})` }}
-                                onMouseEnter={handlePreviewOpen}
-                                onMouseLeave={handlePreviewClose}
-                                data-card-url={s3CardImageURL(match.player1Base, locale)}
-                            />
-                        </Box>
-                        <Box sx={{ ...styles.parentBoxStyling,left:'-15px',top:'24px' }}>
-                            <Box 
-                                sx={{ ...styles.leaderStyleCard,backgroundImage:`url(${s3CardImageURL(match.player1Leader, locale, LeaderBaseCardStyle.PlainLeader)})` }}
-                                onMouseEnter={handlePreviewOpen}
-                                onMouseLeave={handlePreviewClose}
-                                data-card-url={s3CardImageURL(match.player1Leader, locale, LeaderBaseCardStyle.PlainLeader)}
-                            />
-                        </Box>
-                    </Box>
-                    <Typography variant="body1" sx={styles.matchType}>vs</Typography>
-                    <Box sx={{ position:'relative' }}>
-                        <Box>
-                            <Box 
-                                sx={{ ...styles.leaderStyleCard,backgroundImage:`url(${s3CardImageURL(match.player2Base, locale)})` }}
-                                onMouseEnter={handlePreviewOpen}
-                                onMouseLeave={handlePreviewClose}
-                                data-card-url={s3CardImageURL(match.player2Base, locale)}
-                            />
-                        </Box>
-                        <Box sx={{ ...styles.parentBoxStyling,left:'-15px',top:'24px' }}>
-                            <Box 
-                                sx={{ ...styles.leaderStyleCard,backgroundImage:`url(${s3CardImageURL(match.player2Leader, locale, LeaderBaseCardStyle.PlainLeader)})` }}
-                                onMouseEnter={handlePreviewOpen}
-                                onMouseLeave={handlePreviewClose}
-                                data-card-url={s3CardImageURL(match.player2Leader, locale, LeaderBaseCardStyle.PlainLeader)}
-                            />
-                        </Box>
-                    </Box>
+        <Box sx={styles.box}>
+            <Box sx={styles.matchItems}>
+                <Box sx={{ position:'relative', flex: 1 }}>
+                    <OverlappingCards sx={styles.overlappingCard} baseCard={match.player1Base} leaderCard={match.player1Leader} />
                 </Box>
-                {!match.isPrivate && (
-                    <span title={spectateDisabled ? 'Log in to Spectate' : ''}>
-                        <Button
-                            onClick={handleSpectate}
-                            disabled={spectateDisabled}
-                        >
-                            Spectate
-                        </Button>
-                    </span>
-                )}
+                <Typography variant="body1" sx={styles.matchType}>vs</Typography>
+                <Box sx={{ position:'relative', flex: 1 }}>
+                    <OverlappingCards sx={styles.overlappingCard} baseCard={match.player2Base} leaderCard={match.player2Leader} />
+                </Box>
             </Box>
-            <Popover
-                id="mouse-over-popover"
-                sx={{ pointerEvents: 'none' }}
-                open={open}
-                anchorEl={anchorElement}
-                onClose={handlePreviewClose}
-                disableRestoreFocus
-                slotProps={{ paper: { sx: { backgroundColor: 'transparent' } } }}
-                {...popoverConfig()}
-            >
-                <Box sx={{ ...styles.cardPopover, backgroundImage: previewImage }} />
-            </Popover>
-        </>
+            {!match.isPrivate && (
+                <span title={spectateDisabled ? 'Log in to Spectate' : ''}>
+                    <Button
+                        onClick={handleSpectate}
+                        disabled={spectateDisabled}
+                    >
+                        Spectate
+                    </Button>
+                </span>
+            )}
+        </Box>
     );
 };
